requirements.txt
PyYAML==5.1
pycryptodome
pip
XlsxWriter==1.1.5
openpyxl==2.6.1
pyexcel-xls
requests==2.21.0
PyMySQL